Broken: scroll game level on correct note
Issue: since the list of upcoming notes determines where they're drawn,
and notes are removed from the list each time the correct note is
played, AND the viewbox containing these notes is scrolling, it always
skips over one note (which stays out of view).

Fix: need to draw the upcoming notes at the same position! Maybe just
use an array instead of a list of notes, and don't delete them as the
game progresses?